Fixed 64bit build (thanks to jpv950)

+++ src/libs/common/Format.cpp
@@ -429,9 +429,9 @@ CFormat& CFormat::operator%(void * value)
 		// - Linux:   leading zeros missing 
 		// -> format it as hex
 		if (sizeof (void *) == 8) { // 64 bit
-			SetCurrentField(wxString::Format(wxT("0x%016x"), (uint64_t) value));
+			SetCurrentField(wxString::Format(wxT("0x%016x"), (uintptr_t) value));
 		} else { // 32 bit
-			SetCurrentField(wxString::Format(wxT("0x%08x"), (uint32_t) value));
+			SetCurrentField(wxString::Format(wxT("0x%08x"),  (uintptr_t) value));
 		}
 	}
